roadpulse-backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
motor==3.3.1
python-jose>=3.4.0
passlib==1.7.4
python-multipart>=0.0.18
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic>=2.6.0
pymongo==4.5.0 
mapbox==0.18.0
uvicorn[standard]>=0.27.0
Pillow>=10.0.0
torch==2.8.0
torchvision==0.23.0
pandas==2.3.2
scikit-learn ==1.7.1
haversine==2.9.0
googlemaps==4.10.0
polyline==2.0.3
joblib==1.5.1
Pillow>=10.0.0